update help nav through chapter 7
