+++ content/resources/chapter-4/index.ts
@@ -1,11 +1,13 @@
 import * as PublicKeyResourcesThree from './public-key-3'
 import * as PublicKeyResourcesFour from './public-key-4'
-import * as AddressResources from './address'
+import * as AddressResourcesTwo from './address-2'
+import * as AddressResourcesThree from './address-3'
 
 const Chapter4 = {
   'public-key-3': PublicKeyResourcesThree,
   'public-key-4': PublicKeyResourcesFour,
-  address: AddressResources,
+  'address-2': AddressResourcesTwo,
+  'address-3': AddressResourcesThree,
 }
 
 export default Chapter4
