+++ content/resources/chapter-5/derive-message-3.tsx
@@ -1,0 +1,55 @@
+'use client'
+
+import { useTranslations } from 'hooks'
+import { useState } from 'react'
+
+import { Text, ResourcePage, ToggleSwitch, CodeExample } from 'ui'
+import { useAtom } from 'jotai'
+import { currentLanguageAtom } from 'state/state'
+
+export default function DeriveMessageResourcesThree({ lang }) {
+  const t = useTranslations(lang)
+  const [challengeIsToggled, setChallengeIsToggled] = useState(false)
+
+  const challengeToggleSwitch = () => {
+    setChallengeIsToggled(!challengeIsToggled)
+  }
+
+  return (
+    <ResourcePage
+      lang={lang}
+      readingResources={
+        <>
+          <Text className="mt-[25px] text-xl font-bold">
+            {t('chapter_five.resources.derive_message.op_pushdata_heading')}
+          </Text>
+          <Text>
+            {t(
+              'chapter_five.resources.derive_message.op_pushdata_paragraph_one'
+            )}
+          </Text>
+        </>
+      }
+      codeResources={
+        <>
+          <Text>{t('help_page.solution_one')}</Text>
+          <div className="flex flex-row items-center gap-2">
+            <ToggleSwitch
+              checked={challengeIsToggled}
+              onChange={challengeToggleSwitch}
+            />
+            <Text>{t('help_page.spoilers_confirm')}</Text>
+          </div>
+          {challengeIsToggled && (
+            <CodeExample
+              copy
+              language="bash"
+              className="text-wrap text-white"
+              code="304402204e45e16932b8af514961a1d3a1a25fdf3f4f7732e9d624c6c61548ab5fb8cd410220181522ec8eca07de4860a4acdd12909d831cc56cbbac4622082221a8768d1d0901"
+            />
+          )}
+        </>
+      }
+    />
+  )
+}
