+++ content/resources/chapter-5/index.ts
@@ -1,11 +1,29 @@
-import * as VerifySignatureResources from './verify-signature'
-import * as DeriveMessageResources from './derive-message'
-import * as ValidateSignatureResources from './validate-signature'
+import * as VerifySignatureResourcesTwo from './verify-signature-2'
+import * as VerifySignatureResourcesThree from './verify-signature-3'
+import * as VerifySignatureResourcesFour from './verify-signature-4'
+import * as VerifySignatureResourcesFive from './verify-signature-5'
+import * as DeriveMessageResourcesThree from './derive-message-3'
+import * as DeriveMessageResourcesFour from './derive-message-4'
+import * as DeriveMessageResourcesSix from './derive-message-6'
+import * as DeriveMessageResourcesSeven from './derive-message-7'
+import * as ValidateSignatureResourcesOne from './validate-signature-1'
+import * as ValidateSignatureResourcesTwo from './validate-signature-2'
+import * as ValidateSignatureResourcesThree from './validate-signature-3'
+import * as ValidateSignatureResourcesFour from './validate-signature-4'
 
 const Chapter5 = {
-  'derive-message': DeriveMessageResources,
-  'verify-signature': VerifySignatureResources,
-  'validate-signature': ValidateSignatureResources,
+  'derive-message-3': DeriveMessageResourcesThree,
+  'derive-message-4': DeriveMessageResourcesFour,
+  'derive-message-6': DeriveMessageResourcesSix,
+  'derive-message-7': DeriveMessageResourcesSeven,
+  'verify-signature-2': VerifySignatureResourcesTwo,
+  'verify-signature-3': VerifySignatureResourcesThree,
+  'verify-signature-4': VerifySignatureResourcesFour,
+  'verify-signature-5': VerifySignatureResourcesFive,
+  'validate-signature-1': ValidateSignatureResourcesOne,
+  'validate-signature-2': ValidateSignatureResourcesTwo,
+  'validate-signature-3': ValidateSignatureResourcesThree,
+  'validate-signature-4': ValidateSignatureResourcesFour,
 }
 
 export default Chapter5
