+++ content/resources/chapter-5/validate-signature-4.tsx
@@ -1,0 +1,161 @@
+'use client'
+
+import { useTranslations } from 'hooks'
+import { useState } from 'react'
+import { Loader } from 'shared'
+
+import MonacoEditor from '@monaco-editor/react'
+
+import { EditorConfig } from 'types'
+import { Text, ResourcePage, ToggleSwitch } from 'ui'
+import LanguageTabs from 'ui/lesson/ScriptingChallenge/LanguageTabs'
+import { readOnlyOptions } from 'ui/lesson/ScriptingChallenge/config'
+import { getLanguageString } from 'lib/SavedCode'
+import { useAtom } from 'jotai'
+import { currentLanguageAtom } from 'state/state'
+
+const javascriptChallenge = {
+  program: `console.log("KILL")`,
+  defaultFunction: {
+    name: 'verify',
+    args: [],
+  },
+  defaultCode: `function verify_keys(keys) {
+  for (let i = 0; i < keys.length; i++) {
+    let value = keys[i];
+    let valueSub = value.substring(2)
+    let valueX = '0x' + valueSub.substring(0, 64)
+    let valueY = '0x' + valueSub.substring(64)
+    let keyGE = new GE(new FE(valueX), new FE(valueY))
+    // If the public key can verify that means this was the one used to sign
+    if (verify(sig_r_fe, sig_s_fe, keyGE, msg_fe)) {
+      return value
+    }
+  }
+}`,
+  validate: async () => {
+    return [true, undefined]
+  },
+  constraints: [],
+}
+
+const pythonChallenge = {
+  program: `print("KILL")`,
+  defaultFunction: {
+    name: 'verify',
+    args: [],
+  },
+  defaultCode: `def verify_keys(keys):
+    for key in keys:
+        key_trim = key[2:len(key)]
+        key_x = int(key_trim[0:64], 16)
+        key_y = int(key_trim[64:len(key_trim)], 16)
+        key_ge = GE(key_x, key_y)
+
+        if verify(sig_r, sig_s, key_ge, msg):
+            return key`,
+  validate: async () => {
+    return [true, undefined]
+  },
+  constraints: [],
+}
+
+const config: EditorConfig = {
+  defaultLanguage: 'javascript',
+  languages: {
+    javascript: javascriptChallenge,
+    python: pythonChallenge,
+  },
+}
+
+export default function VerifySignatureResources({ lang }) {
+  const t = useTranslations(lang)
+  const [currentLanguage] = useAtom(currentLanguageAtom)
+
+  const initialStateCode =
+    config.languages[getLanguageString(currentLanguage)].defaultCode
+  const [code, setCode] = useState(initialStateCode as string)
+
+  const [language, setLanguage] = useState(getLanguageString(currentLanguage))
+
+  const [challengeIsToggled, setChallengeIsToggled] = useState(false)
+
+  const challengeToggleSwitch = () => {
+    setChallengeIsToggled(!challengeIsToggled)
+  }
+
+  const handleSetLanguage = (value) => {
+    setLanguage(value)
+    setCode(config.languages[value].defaultCode as string)
+  }
+
+  const handleBeforeMount = (monaco) => {
+    monaco.editor.defineTheme('satoshi', {
+      base: readOnlyOptions.theme,
+      inherit: true,
+      rules: [],
+      colors: {
+        'editor.background': '#00000000',
+        'editor.lineHighlightBorder': '#00000000', // 4th channel is for transparency
+      },
+    })
+  }
+
+  const handleMount = (_editor, monaco) => {
+    monaco.editor.setTheme('satoshi')
+  }
+
+  return (
+    <ResourcePage
+      lang={lang}
+      readingResources={
+        <>
+          <Text className="mt-[25px] text-xl font-bold">
+            {t(
+              'chapter_five.resources.validate_signature.message_verification_heading'
+            )}
+          </Text>
+          <Text>
+            {t(
+              'chapter_five.resources.validate_signature.message_verification_paragraph_one'
+            )}
+          </Text>
+        </>
+      }
+      codeResources={
+        <>
+          <Text>{t('help_page.solution_one')}</Text>
+          <div className="flex flex-row items-center gap-2">
+            <ToggleSwitch
+              checked={challengeIsToggled}
+              onChange={challengeToggleSwitch}
+            />
+            <Text>{t('help_page.spoilers_confirm')}</Text>
+          </div>
+          {challengeIsToggled && (
+            <div className="border border-white/25">
+              <LanguageTabs
+                languages={config.languages}
+                value={language}
+                onChange={handleSetLanguage}
+                noHide
+              />
+              <div className="relative grow bg-[#00000026] font-mono text-sm text-white">
+                <MonacoEditor
+                  loading={<Loader className="h-10 w-10 text-white" />}
+                  height={`255px`}
+                  value={code}
+                  beforeMount={handleBeforeMount}
+                  onMount={handleMount}
+                  language={language}
+                  theme={'satoshi'}
+                  options={readOnlyOptions}
+                />
+              </div>
+            </div>
+          )}
+        </>
+      }
+    />
+  )
+}
