+++ content/resources/chapter-6/in-out-5.tsx
@@ -1,0 +1,175 @@
+'use client'
+
+import { useTranslations } from 'hooks'
+import { useState } from 'react'
+import { Loader } from 'shared'
+
+import MonacoEditor from '@monaco-editor/react'
+
+import { EditorConfig } from 'types'
+import { Text, ResourcePage, ToggleSwitch } from 'ui'
+import LanguageTabs from 'ui/lesson/ScriptingChallenge/LanguageTabs'
+import { readOnlyOptions } from 'ui/lesson/ScriptingChallenge/config'
+import { getLanguageString } from 'lib/SavedCode'
+import { useAtom } from 'jotai'
+import { currentLanguageAtom } from 'state/state'
+
+const javascriptChallenge = {
+  program: `console.log("KILL")`,
+  defaultFunction: {
+    name: 'verify',
+    args: [],
+  },
+  defaultCode: `class Output {
+  constructor() {
+    this.value = 0;
+    this.witness_version = 0;
+    this.witness_data = Buffer.alloc(0);
+  }
+
+  static from_options(addr, value) {
+    assert(Number.isInteger(value));
+    const self = new this();
+    const {version, program} = bech32.decode('bc', addr);
+    self.witness_version = version;
+    self.witness_data = Buffer.from(program);
+    self.value = value;
+    return self;
+  }
+
+  serialize() {
+    const buf = Buffer.alloc(11);
+    buf.writeBigInt64LE(BigInt(this.value), 0);
+    buf.writeUInt8(this.witness_data.length + 2, 8);
+    buf.writeUInt8(this.witness_version, 9);
+    buf.writeUInt8(this.witness_data.length, 10);
+    return Buffer.concat([buf, this.witness_data]);
+  }
+}`,
+  validate: async () => {
+    return [true, undefined]
+  },
+  constraints: [],
+}
+
+const pythonChallenge = {
+  program: `print("KILL")`,
+  defaultFunction: {
+    name: 'verify',
+    args: [],
+  },
+  defaultCode: `class Output:
+    def __init__(self):
+        self.value = 0
+        self.witness_version = 0
+        self.witness_data = b""
+
+    @classmethod
+    def from_options(cls, addr, value):
+        assert isinstance(value, int)
+        self = cls()
+        (ver, data) = bech32.decode("bc", addr)
+        self.witness_version = ver
+        self.witness_data = bytes(data)
+        self.value = value
+        return self
+
+    def serialize(self):
+        r = b""
+        r += pack("<q", self.value)
+        r += pack("<B", len(self.witness_data) + 2)
+        r += pack("<B", self.witness_version)
+        r += pack("<B", len(self.witness_data))
+        r += self.witness_data
+        return r`,
+  validate: async () => {
+    return [true, undefined]
+  },
+  constraints: [],
+}
+
+const config: EditorConfig = {
+  defaultLanguage: 'javascript',
+  languages: {
+    javascript: javascriptChallenge,
+    python: pythonChallenge,
+  },
+}
+
+export default function InOutResourcesFive({ lang }) {
+  const t = useTranslations(lang)
+  const [currentLanguage] = useAtom(currentLanguageAtom)
+  const initialStateCode =
+    config.languages[getLanguageString(currentLanguage)].defaultCode
+  const [code, setCode] = useState(initialStateCode as string)
+
+  const [language, setLanguage] = useState(getLanguageString(currentLanguage))
+
+  const [challengeIsToggled, setChallengeIsToggled] = useState(false)
+
+  const challengeToggleSwitch = () => {
+    setChallengeIsToggled(!challengeIsToggled)
+  }
+
+  const handleSetLanguage = (value) => {
+    setLanguage(value)
+    setCode(config.languages[value].defaultCode as string)
+  }
+
+  const handleBeforeMount = (monaco) => {
+    monaco.editor.defineTheme('satoshi', {
+      base: readOnlyOptions.theme,
+      inherit: true,
+      rules: [],
+      colors: {
+        'editor.background': '#00000000',
+        'editor.lineHighlightBorder': '#00000000', // 4th channel is for transparency
+      },
+    })
+  }
+
+  const handleMount = (_editor, monaco) => {
+    monaco.editor.setTheme('satoshi')
+  }
+
+  return (
+    <ResourcePage
+      lang={lang}
+      readingResources={<></>}
+      codeResources={
+        <>
+          <Text>{t('help_page.solution_one')}</Text>
+          <div className="flex flex-row items-center gap-2">
+            <ToggleSwitch
+              checked={challengeIsToggled}
+              onChange={challengeToggleSwitch}
+            />
+            <Text>{t('help_page.spoilers_confirm')}</Text>
+          </div>
+          {challengeIsToggled && (
+            <div className="border border-white/25">
+              <LanguageTabs
+                languages={config.languages}
+                value={language}
+                onChange={handleSetLanguage}
+                noHide
+              />
+              <div className="relative grow bg-[#00000026] font-mono text-sm text-white">
+                <MonacoEditor
+                  loading={<Loader className="h-10 w-10 text-white" />}
+                  height={`505px`}
+                  value={code}
+                  beforeMount={handleBeforeMount}
+                  onMount={handleMount}
+                  language={language}
+                  theme={'satoshi'}
+                  options={readOnlyOptions}
+                />
+              </div>
+            </div>
+          )}
+        </>
+      }
+    />
+  )
+}
