+++ content/resources/chapter-6/index.ts
@@ -1,9 +1,13 @@
-import * as InOutResources from './in-out'
-import * as PutItTogetherResources from './put-it-together'
+import * as InOutResourcesFive from './in-out-5'
+import * as PutItTogetherResourcesOne from './put-it-together-1'
+import * as PutItTogetherResourcesTwo from './put-it-together-2'
+import * as PutItTogetherResourcesThree from './put-it-together-3'
 
 const Chapter6 = {
-  'in-out': InOutResources,
-  'put-it-together': PutItTogetherResources,
+  'in-out-5': InOutResourcesFive,
+  'put-it-together-1': PutItTogetherResourcesOne,
+  'put-it-together-2': PutItTogetherResourcesTwo,
+  'put-it-together-3': PutItTogetherResourcesThree,
 }
 
 export default Chapter6
